XBE: Fix responding to messages for other devices
In case of a device ID mismatch, the previous logic would evaluate the
function ID next and respond with Status::IllegalFunction. We cannot
make assumptions about function ID if the message is not for us.

Separate dispatch from the callback calling, retrieve the device from
the dispatch early and ignore the message if not for us.

BFW-7755