Bigger buffers in phase stepping
This prevents starving events in phase stepping in case of high-speed
curved moves, where the input shaper produces insane amounts of very
very tiny micro moves. This in turn prevents layer shifts.

This is probably just a workaround. The input shaper (running in the
stepper interrupt, which is very high priority) should have enough time
to refill the buffers. It seems that once it fills the output buffer, it
doesn't get called again when it empties and instead waits for some
other event on other axis (Z or E). This needs some further
investigation, but for now this seems to improve the situation
significantly.

The new value of 128 was discovered by experimental means with a very
hard-core layer shift reproducer, so it should be safe for any practical
means.

BFW-7873.