emergency_stop: Don't BSOD during probing
Probing/homing moves have a bit of a special place, because the endstops are enabled during them.
Endstops, when triggered, cause quickstops.

So during homing, we don't want to do the "standard" segment-and-maybe-block thing, because weird things would start happening.

Instead of that (and instead of just BSODing), we now do a quick_stop (which is the same as the endstop being triggered) and block after we get out of the probing code.

BFW-7544