C1L: Tune parking position
Let the filament flow unobstructed to the bottom of the printer - avoid spilling it onto the bed.
That allows purging and other messy operations sometimes finish clean and the filament string separated from the nozzle purely by gravity.

Also, remove idiotic parking to Z=168, there is no technical reason for it on C1L.

BFW-8108